feat: changes based on review comments
