+++ test/unit/token/env-change-token-parser-test.js
@@ -1,5 +1,8 @@
 const StreamParser = require('../../../src/token/stream-parser');
 const WritableTrackingBuffer = require('../../../src/tracking-buffer/writable-tracking-buffer');
+const { Connection } = require('../../../src/tedious');
+const Message = require('../../../src/message');
+const Login7TokenHandler = require('../../../src/token/handler').Login7TokenHandler;
 const assert = require('chai').assert;
 
 describe('Env Change Token Parser', () => {
@@ -65,5 +68,30 @@ describe('Env Change Token Parser', () => {
     const parser = StreamParser.parseTokens([data], {}, {});
     const result = await parser.next();
     assert.isTrue(result.done);
+  });
+
+  it('Test if routing data capture the correct instance name value', async function() {
+    const valueBuffer = new WritableTrackingBuffer(0);
+    valueBuffer.writeUInt8(0); // Protocol
+    valueBuffer.writeUInt16LE(1433); // Port
+    valueBuffer.writeUsVarchar('127.0.0.1\\instanceNameA', 'ucs-2');
+
+    const envValueDataBuffer = new WritableTrackingBuffer(0);
+    envValueDataBuffer.writeUInt8(20); // Type
+    envValueDataBuffer.writeUsVarbyte(valueBuffer.data);
+    envValueDataBuffer.writeUsVarbyte(Buffer.alloc(0));
+
+    const envChangeBuffer = new WritableTrackingBuffer(0);
+    envChangeBuffer.writeUInt8(0xE3); // TokenType
+    envChangeBuffer.writeUsVarbyte(envValueDataBuffer.data); // Length + EnvValueData
+
+    const responseMessage = new Message({ type: 0x04 });
+    responseMessage.end(envChangeBuffer.data);
+    const parser = StreamParser.parseTokens(responseMessage, {}, {});
+    const result = await parser.next();
+    const handler = new Login7TokenHandler(new Connection({ server: 'servername' }));
+    handler[result.value.handlerName](result.value);
+    assert.strictEqual(handler.routingData.instanceName, 'instanceNameA');
+    assert.isTrue((await parser.next()).done);
   });
 });
